double filter time

+++ host.c
@@ -552,7 +552,7 @@ Host_FilterTime
 Returns false if the time is too short to run a frame
 ===================
 */
-qboolean Host_FilterTime (float time)
+qboolean Host_FilterTime (double time)
 {
 	float maxfps;
